Added new module for card-builder handling
